+++ src/Request/InputFiles.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace MagicObject\Request;
+
+use MagicObject\File\PicoUploadFile;
+
+/**
+ * Class InptFiles
+ *
+ * This class extends the PicoUploadFile class to handle input files.
+ * It is designed to manage uploaded files and provide methods for file handling.
+ *
+ * @package MagicObject\Request
+ */
+class InptFiles extends PicoUploadFile
+{
+    
+}